refactor: PostService에서 InMemoryRepository 결합 제거
- InMemoryRepository내부에서 Reflection을 사용하여 InMemoryRepository만을 위한 로직들을 제거
