+++ src/main/java/org/sopt/domain/post/repository/PostRepository.java
@@ -1,0 +1,15 @@
+package org.sopt.domain.post.repository;
+
+import org.sopt.domain.post.entity.Post;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface PostRepository {
+
+    Post save(Post post);
+
+    List<Post> findAll();
+    Optional<Post> findById(Long id);
+    boolean deleteById(Long id);
+}
