+++ src/main/java/org/sopt/domain/post/service/PostService.java
@@ -6,7 +6,7 @@ import org.sopt.domain.post.dto.request.CreatePostRequest;
 import org.sopt.domain.post.dto.response.CreatePostResponse;
 import org.sopt.domain.post.dto.response.PostResponse;
 import org.sopt.domain.post.exception.PostNotFoundException;
-import org.sopt.domain.post.repository.InMemoryRepository;
+import org.sopt.domain.post.repository.PostRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -16,27 +16,24 @@ import static org.sopt.domain.post.validator.PostValidator.validateCreatePostReq
 @Service
 @RequiredArgsConstructor
 public class PostService {
-    private final InMemoryRepository postRepository;
+    private final PostRepository postRepository;
 
-    private Long nextId = 1L;
-
-    // CREATE ✅ 같이 구현
-    // 글쓰기 화면에서 "완료" 버튼을 누르면 이 메서드가 호출돼요
+    // CREATE
     public CreatePostResponse createPost(CreatePostRequest request) {
 
         validateCreatePostRequest(request.title(), request.content());
 
         String createdAt = java.time.LocalDateTime.now().toString();
-        Post post = new Post(nextId++, request.title(), request.content(), request.author(), createdAt);
 
-        postRepository.save(post);
-        return new CreatePostResponse(post.getId());
+        Post post = new Post(request.title(), request.content(), request.author(), createdAt);
+
+        Post savedPost = postRepository.save(post);
+
+        return new CreatePostResponse(savedPost.getId());
     }
 
 
-
-    // READ - 전체 📝 과제
-    // 자유게시판 목록 화면에서 호출돼요
+    // READ - 전체 목록
     public List<PostResponse> getAllPosts() {
 
         List<Post> posts = postRepository.findAll();
@@ -46,15 +43,13 @@ public class PostService {
                 .toList();
     }
 
-    // READ - 단건 📝 과제
-    // 목록에서 특정 게시글을 탭하면 호출돼요 (게시글 상세 화면)
+    // READ - 상세
     public PostResponse getPost(Long id) {
         Post post = postRepository.findById(id)
                  .orElseThrow(() -> new PostNotFoundException(id));
             return new PostResponse(post);
     }
-    // UPDATE 📝 과제
-    // 게시글 수정 화면에서 "완료"를 누르면 호출돼요
+    // UPDATE
     public void updatePost(Long id, String newTitle, String newContent) {
 
         Post post = postRepository.findById(id)
@@ -65,8 +60,7 @@ public class PostService {
         post.update(newTitle, newContent);
     }
 
-    // DELETE 📝 과제
-    // 게시글 상세에서 삭제를 누르면 호출돼요
+    // DELETE
     public void deletePost(Long id) {
 
         boolean deleted = postRepository.deleteById(id);
